requirements.txt
diffusers
gradio
huggingface_hub
numpy
peft
Pillow
torch
torchvision
transformers
